+++ src/RestApiImpl.cpp
@@ -48,7 +48,7 @@ namespace Huobi {
             std::vector<Candlestick> cans;
             JsonWrapper data = json.getJsonObjectOrArray("data");
 
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 Candlestick candlestick;
                 candlestick.timestamp = item.getLong("id");
@@ -80,10 +80,10 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             std::vector<Trade> trades;
             JsonWrapper dataArray = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < dataArray.size(); i++) {
+            for (size_t i = 0; i < dataArray.size(); i++) {
                 JsonWrapper item = dataArray.getJsonObjectAt(i);
                 JsonWrapper dataArrayIn = item.getJsonObjectOrArray("data");
-                for (int j = 0; j < dataArrayIn.size(); j++) {
+                for (size_t j = 0; j < dataArrayIn.size(); j++) {
                     JsonWrapper itemIn = dataArrayIn.getJsonObjectAt(j);
                     Trade trade;
                     trade.price = itemIn.getDecimal("price");
@@ -107,7 +107,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             std::vector<Symbols> symbolsList;
             JsonWrapper data = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 Symbols symbols;
                 symbols.baseCurrency = item.getString("base-currency");
@@ -134,7 +134,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             std::vector<std::string> stringList;
             JsonWrapper dataArray = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < dataArray.size(); i++) {
+            for (size_t i = 0; i < dataArray.size(); i++) {
                 stringList.push_back(dataArray.getStringAt(i));
             }
 
@@ -170,7 +170,7 @@ namespace Huobi {
             std::vector<Account>accounts;
             JsonWrapper data = json.getJsonObjectOrArray("data");
             size_t size = data.size();
-            for (int i = 0; i < size; i++) {
+            for (size_t i = 0; i < size; i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 Account account;
                 account.id = item.getLong("id");
@@ -193,7 +193,7 @@ namespace Huobi {
             std::vector<Balance> balances;
             JsonWrapper data = json.getJsonObjectOrArray("data");
             JsonWrapper list = data.getJsonObjectOrArray("list");
-            for (int i = 0; i < list.size(); i++) {
+            for (size_t i = 0; i < list.size(); i++) {
                 JsonWrapper item = list.getJsonObjectAt(i);
                 Balance balance;
                 balance.balance = item.getDecimal("balance");
@@ -260,7 +260,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             std::vector<Withdraw> withdraws;
             JsonWrapper data = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 Withdraw withdraw;
                 withdraw.id = item.getLong("id");
@@ -296,7 +296,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             std::vector<Deposit> lstdeposit;
             JsonWrapper data = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 Deposit deposit;
                 deposit.id = item.getLong("id");
@@ -416,7 +416,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             std::vector<Loan> loans;
             JsonWrapper data = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 Loan loan;
                 loan.loanBalance = item.getDecimal("loan-balance");
@@ -466,7 +466,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             std::vector<Order> orderList;
             JsonWrapper data = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 Order order;
                 order.orderId = item.getLong("id");
@@ -575,10 +575,10 @@ namespace Huobi {
             JsonWrapper data = json.getJsonObjectOrArray("data");
             JsonWrapper success = data.getJsonObjectOrArray("success");
             JsonWrapper failed = data.getJsonObjectOrArray("failed");
-            for (int i = 0; i < success.size(); i++) {
+            for (size_t i = 0; i < success.size(); i++) {
                 batchCancelOrdersResult.successList.push_back(success.getStringAt(i));
             }
-            for (int i = 0; i < failed.size(); i++) {
+            for (size_t i = 0; i < failed.size(); i++) {
                 JsonWrapper item = failed.getJsonObjectAt(i);
                 FailedObj failedObj;
                 failedObj.errMsg = item.getString("err-msg");
@@ -637,7 +637,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             std::vector<MatchResult> matchResultList;
             JsonWrapper data = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 MatchResult matchResult;
                 matchResult.id = (item.getLong("id"));
@@ -679,7 +679,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             std::vector<MatchResult> matchResultList;
             JsonWrapper data = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 MatchResult matchResult;
                 matchResult.id = item.getLong("id");
@@ -778,7 +778,7 @@ namespace Huobi {
         res->jsonParser = [this] (const JsonWrapper & json) {
             std::vector<Order> orderList;
             JsonWrapper data = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 Order order;
                 order.accountType =
@@ -827,7 +827,7 @@ namespace Huobi {
         res->jsonParser = [this](JsonWrapper json) {
             std::vector<Balance> balances;
             JsonWrapper data = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 Balance balance;
                 balance.currency = item.getString("currency");
@@ -850,14 +850,14 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             std::vector<CompleteSubAccountInfo> completeSubAccountInfos;
             JsonWrapper data = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 CompleteSubAccountInfo completeSubAccountInfo;
                 completeSubAccountInfo.id = item.getLong("id");
                 completeSubAccountInfo.type = AccountType::lookup(item.getString("type"));
                 JsonWrapper list = item.getJsonObjectOrArray("list");
                 std::vector<Balance> balances;
-                for (int j = 0; j < list.size(); j++) {
+                for (size_t j = 0; j < list.size(); j++) {
                     JsonWrapper in = list.getJsonObjectAt(j);
                     Balance balance;
                     balance.currency = in.getString("currency");
@@ -891,7 +891,7 @@ namespace Huobi {
                     EtfStatus::lookup(std::to_string(data.getInt("etf_status")));
             JsonWrapper unitPrices = data.getJsonObjectOrArray("unit_price");
             std::vector<UnitPrice> unitPriceList;
-            for (int i = 0; i < unitPrices.size(); i++) {
+            for (size_t i = 0; i < unitPrices.size(); i++) {
                 JsonWrapper item = unitPrices.getJsonObjectAt(i);
                 UnitPrice unitPrice;
                 unitPrice.currency = item.getString("currency");
@@ -938,7 +938,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             std::vector<EtfSwapHistory> etfSwapHistoryList;
             JsonWrapper data = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 EtfSwapHistory etfSwapHistory;
                 etfSwapHistory.createdTimestamp = item.getLong("gmt_created");
@@ -952,7 +952,7 @@ namespace Huobi {
                 etfSwapHistory.pointCardAmount = detail.getDecimal("point_card_amount");
                 JsonWrapper usedCurrencyArray = detail.getJsonObjectOrArray("used_currency_list");
                 std::vector<UnitPrice> usedCurrencyList;
-                for (int j = 0; j < usedCurrencyArray.size(); j++) {
+                for (size_t j = 0; j < usedCurrencyArray.size(); j++) {
                     JsonWrapper currency = usedCurrencyArray.getJsonObjectAt(i);
                     UnitPrice unitPrice;
                     unitPrice.amount = currency.getDecimal("amount");
@@ -962,7 +962,7 @@ namespace Huobi {
                 etfSwapHistory.usedCurrencyList = usedCurrencyList;
                 JsonWrapper obtainCurrencyArray = detail.getJsonObjectOrArray("obtain_currency_list");
                 std::vector<UnitPrice> obtainCurrencyList;
-                for (int j = 0; j < obtainCurrencyArray.size(); j++) {
+                for (size_t j = 0; j < obtainCurrencyArray.size(); j++) {
                     JsonWrapper currency = obtainCurrencyArray.getJsonObjectAt(j);
                     UnitPrice unitPrice;
                     unitPrice.amount = currency.getDecimal("amount");
@@ -992,7 +992,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             std::vector<Candlestick> cands;
             JsonWrapper data = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 Candlestick candlestick;
                 candlestick.timestamp = item.getLong("id");
@@ -1023,7 +1023,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             std::vector<MarginBalanceDetail> marginBalanceDetailList;
             JsonWrapper dataArray = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < dataArray.size(); i++) {
+            for (size_t i = 0; i < dataArray.size(); i++) {
                 JsonWrapper itemInData = dataArray.getJsonObjectAt(i);
                 MarginBalanceDetail marginBalanceDetail;
                 marginBalanceDetail.id = itemInData.getLong("id");
@@ -1034,7 +1034,7 @@ namespace Huobi {
                 marginBalanceDetail.state = AccountState::lookup(itemInData.getString("state"));
                 marginBalanceDetail.riskRate = itemInData.getDecimal("risk-rate");
                 JsonWrapper listArray = itemInData.getJsonObjectOrArray("list");
-                for (int i = 0; i < listArray.size(); i++) {
+                for (size_t i = 0; i < listArray.size(); i++) {
                     JsonWrapper itemInList = listArray.getJsonObjectAt(i);
                     Balance balance;
                     balance.currency = itemInList.getString("currency");
@@ -1103,7 +1103,7 @@ namespace Huobi {
             std::vector<FeeRate> feeRatelList;
             JsonWrapper dataArray = json.getJsonObjectOrArray("data");
 
-            for (int i = 0; i < dataArray.size(); i++) {
+            for (size_t i = 0; i < dataArray.size(); i++) {
                 JsonWrapper itemInData = dataArray.getJsonObjectAt(i);
                 FeeRate feeRate;
                 feeRate.symbol = itemInData.getString("symbol");
@@ -1148,7 +1148,7 @@ namespace Huobi {
         res->jsonParser = [this] (const JsonWrapper & json) {
             std::vector<Order> orderList;
             JsonWrapper data = json.getJsonObjectOrArray("data");
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 Order order;
                 order.accountType =
@@ -1211,14 +1211,14 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             JsonWrapper data = json.getJsonObjectOrArray("data");
             std::vector<CurrencyChain> currencyChains;
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper obj = data.getJsonObjectAt(i);
                 CurrencyChain currencyChain;
                 currencyChain.currency = obj.getString("currency");
                 currencyChain.instStatus = obj.getString("instStatus");
                 JsonWrapper jsonChains = obj.getJsonObjectOrArray("chains");
                 std::vector<Chain> chains;
-                for (int i = 0; i < jsonChains.size(); i++) {
+                for (size_t i = 0; i < jsonChains.size(); i++) {
                     JsonWrapper item = jsonChains.getJsonObjectAt(i);
                     Chain chain;
                     chain.chain = item.getString("chain");
@@ -1265,7 +1265,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             JsonWrapper data = json.getJsonObjectOrArray("data");
             std::vector<DepositAddress> depositAddressVec;
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 DepositAddress depositAddress;
                 depositAddress.address = item.getString("address");
@@ -1296,7 +1296,7 @@ namespace Huobi {
 
             JsonWrapper chains = data.getJsonObjectOrArray("chains");
             std::vector<WithdrawChainQuota> withdrawChainQuotaVec;
-            for (int i = 0; i < chains.size(); i++) {
+            for (size_t i = 0; i < chains.size(); i++) {
                 JsonWrapper item = chains.getJsonObjectAt(i);
                 WithdrawChainQuota withdrawChainQuota;
                 withdrawChainQuota.chain = item.getString("chain");
@@ -1343,7 +1343,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             JsonWrapper data = json.getJsonObjectOrArray("data");
             std::vector<AccountHistory> accountHistories;
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 AccountHistory accountHistory;
                 accountHistory.accountId = item.getLong("account-id");
@@ -1455,7 +1455,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             JsonWrapper data = json.getJsonObjectOrArray("data");
             std::vector<CrossMarginLoadOrder> crossMarginLoadOrders;
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 CrossMarginLoadOrder crossMarginLoadOrder;
                 crossMarginLoadOrder.currency = item.getString("currency");
@@ -1495,7 +1495,7 @@ namespace Huobi {
             crossMarginAccount.type = AccountType::lookup(item.getString("type"));
             JsonWrapper list = item.getJsonObjectOrArray("list");
             std::vector<Balance> balances;
-            for (int i = 0; i < list.size(); i++) {
+            for (size_t i = 0; i < list.size(); i++) {
                 JsonWrapper jw = list.getJsonObjectAt(i);
                 Balance balance;
                 balance.balance = jw.getDecimal("balance");
@@ -1570,7 +1570,7 @@ namespace Huobi {
         res->jsonParser = [this](const JsonWrapper & json) {
             JsonWrapper data = json.getJsonObjectOrArray("data");
             std::vector<BatchOrderResult> batchOrderResults;
-            for (int i = 0; i < data.size(); i++) {
+            for (size_t i = 0; i < data.size(); i++) {
                 JsonWrapper item = data.getJsonObjectAt(i);
                 BatchOrderResult batchOrderResult;
                 batchOrderResult.orderId = item.getLong("order-id");
